Use dedicated headers/trailer types
