+++ examples/access_logger/main_test.go
@@ -17,7 +17,7 @@ func TestHelloWorld_OnTick(t *testing.T) {
 	host := proxytest.NewHostEmulator(opt)
 	defer host.Done() // release the host emulation lock so that other test cases can insert their own host emulation
 
-	host.CallOnLogForAccessLogger([][2]string{{":path", "/this/is/path"}}, nil) // call OnLog
+	host.CallOnLogForAccessLogger(types.Headers{{":path", "/this/is/path"}}, nil) // call OnLog
 
 	logs := host.GetLogs(types.LogLevelInfo)
 	require.Greater(t, len(logs), 0)
